python/requirements_test.txt
# build/test only requirements: not installed in Docker image
black==22.12.0
grpc-stubs==1.24.11  # mypy types for grpcio? TODO: test/verify?
grpcio-tools==1.51.1
mypy==0.991
pytest==7.2.0
types-protobuf==4.21.0.2  # mypy types for protobuf
